Move flexible buffers to crate level
