+++ fidget-wgpu/src/lib.rs
@@ -2,10 +2,13 @@
 #![warn(missing_docs)]
 pub mod voxel;
 
+use fidget_core::render::ImageSize;
 use heck::ToShoutySnakeCase;
 
 /// Re-export the [`wgpu`] module
 pub use wgpu;
+
+////////////////////////////////////////////////////////////////////////////////
 
 /// Returns a set of constant definitions for each opcode
 fn opcode_constants() -> String {
@@ -15,6 +18,8 @@ fn opcode_constants() -> String {
     }
     out
 }
+
+////////////////////////////////////////////////////////////////////////////////
 
 /// Error type for [`init`]
 #[derive(Debug, thiserror::Error)]
@@ -51,4 +56,250 @@ pub async fn init() -> Result<(wgpu::Device, wgpu::Queue), InitError> {
         })
         .await?;
     Ok(out)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+
+/// Handle around a growable GPU buffer which pretends to be smaller
+struct GenericFlexBuffer<T, B, const U: u32> {
+    /// Current item count, which may be smaller than the buffer's capacity
+    item_count: usize,
+    /// Actual GPU buffer
+    data: wgpu::Buffer,
+    /// Buffer label (to be used when reallocating)
+    name: String,
+    /// Marker for buffer data type
+    _t: std::marker::PhantomData<T>,
+    /// Marker for buffer size type
+    _b: std::marker::PhantomData<B>,
+}
+
+/// Flexible buffer which can be resized with a single item count
+type ArrayBuffer<T, const U: u32> = GenericFlexBuffer<T, usize, U>;
+
+/// Flexible buffer which can be resized to fit an image size
+type ImageBuffer<T, const U: u32> = GenericFlexBuffer<T, ImageSize, U>;
+
+/// Module containing usage constants
+mod usage {
+    /// Helper macro to generate usage constants
+    macro_rules! u {
+        ($($flag:ident),+ $(,)?) => {
+            $( wgpu::BufferUsages::$flag.bits() )|+
+        };
+    }
+    pub const STORAGE_COPY_DST: u32 = u!(STORAGE, COPY_DST);
+    pub const STORAGE_INDIRECT: u32 = u!(STORAGE, INDIRECT);
+    pub const STORAGE_INDIRECT_COPY_DST: u32 = u!(STORAGE, INDIRECT, COPY_DST);
+    pub const STORAGE_COPY_SRC_DST: u32 = u!(STORAGE, COPY_SRC, COPY_DST);
+    pub const COPY_DST_MAP_READ: u32 = u!(COPY_DST, MAP_READ);
+}
+
+trait BufferItemCount {
+    fn item_count(&self) -> usize;
+}
+
+impl BufferItemCount for usize {
+    fn item_count(&self) -> usize {
+        *self
+    }
+}
+
+impl BufferItemCount for ImageSize {
+    fn item_count(&self) -> usize {
+        usize::try_from(self.width()).unwrap()
+            * usize::try_from(self.height()).unwrap()
+    }
+}
+
+impl<T, B: BufferItemCount, const U: u32> GenericFlexBuffer<T, B, U> {
+    fn new(
+        device: &wgpu::Device,
+        name: String,
+        item_count: B,
+    ) -> Result<Self, BufferSizeError> {
+        let item_count = item_count.item_count();
+        let size =
+            u64::try_from(item_count * std::mem::size_of::<T>()).unwrap();
+        assert_eq!(size % 4, 0);
+        let usage = wgpu::BufferUsages::from_bits(U).unwrap();
+        Self::check_size(usage, size)?;
+        let data = device.create_buffer(&wgpu::BufferDescriptor {
+            label: Some(name.as_str()),
+            size,
+            usage,
+            mapped_at_creation: false,
+        });
+        Ok(Self {
+            data,
+            item_count,
+            name,
+            _t: std::marker::PhantomData,
+            _b: std::marker::PhantomData,
+        })
+    }
+
+    /// Returns the active buffer size (in bytes)
+    fn size(&self) -> u64 {
+        u64::try_from(self.item_count * std::mem::size_of::<T>()).unwrap()
+    }
+
+    fn check_size(
+        usage: wgpu::BufferUsages,
+        size: u64,
+    ) -> Result<(), BufferSizeError> {
+        let buf_ty = if usage.contains(wgpu::BufferUsages::STORAGE) {
+            BufferType::Storage
+        } else if usage.contains(wgpu::BufferUsages::UNIFORM) {
+            BufferType::Uniform
+        } else {
+            BufferType::Generic
+        };
+        buf_ty.check(size)
+    }
+
+    /// Grows the buffer to fit a particular size in bytes
+    ///
+    /// If the buffer already fits that size, then no allocation is performed,
+    /// but we always update the internal `member_count` (e.g. so that
+    /// [`bind_active`](Self::bind_active) returns the correct subset of the
+    /// buffer).
+    fn grow_to_fit(
+        &mut self,
+        device: &wgpu::Device,
+        item_count: B,
+    ) -> Result<(), BufferSizeError> {
+        let item_count = item_count.item_count();
+        if item_count > self.item_capacity() {
+            let size =
+                u64::try_from(item_count * std::mem::size_of::<T>()).unwrap();
+            assert_eq!(size % 4, 0);
+            let usage = self.data.usage();
+            Self::check_size(usage, size)?;
+            self.data = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(self.name.as_str()),
+                size,
+                usage,
+                mapped_at_creation: false,
+            });
+        }
+        self.item_count = item_count;
+        Ok(())
+    }
+
+    /// Returns a binding resource for the active slice of the buffer
+    fn bind_active(&self) -> wgpu::BindingResource<'_> {
+        self.data.slice(0..self.size()).into()
+    }
+
+    /// Returns the buffer's total capacity (in items)
+    ///
+    /// This may be larger than [`self.item_count`](Self::item_count)
+    fn item_capacity(&self) -> usize {
+        let c = usize::try_from(self.capacity()).unwrap();
+        assert_eq!(c % std::mem::size_of::<T>(), 0);
+        c / std::mem::size_of::<T>()
+    }
+
+    /// Returns the total buffer capacity (in bytes)
+    fn capacity(&self) -> u64 {
+        self.data.size()
+    }
+
+    /// Maps the active portion of the buffer for reading
+    fn map_async(
+        &self,
+        callback: impl FnOnce(Result<(), wgpu::BufferAsyncError>)
+        + wgpu::WasmNotSend
+        + 'static,
+    ) -> wgpu::BufferSlice<'_> {
+        let slice = self.data.slice(0..self.size());
+        slice.map_async(wgpu::MapMode::Read, callback);
+        slice
+    }
+
+    /// Clears the active portion of the buffer
+    fn clear(&self, encoder: &mut wgpu::CommandEncoder) {
+        encoder.clear_buffer(&self.data, 0, Some(self.size()));
+    }
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Error handling zone!  This is perhaps a bit overengineered, but it meets the
+// desired behavior of function error types only containing errors that they can
+// actually return.
+
+/// Error type when resizing a buffer beyond its limit
+///
+/// We check against maximum buffer sizes (from the WebGPU spec) and return an
+/// error immediately, instead of deferring the error to the point where the
+/// buffer is used.
+#[derive(Debug, thiserror::Error)]
+pub enum BufferSizeError {
+    /// Buffer size is not aligned to 4 bytes
+    #[error("requested size {0} must be a multiple of 4 bytes")]
+    NotAligned(u64),
+
+    /// Buffer size is too large for the requested buffer usage
+    #[error(
+        "requested size {requested_size} exceeds maximum {} for \
+        {buffer_type} buffer",
+        buffer_type.max_size()
+    )]
+    TooLarge {
+        /// Size requested (in bytes)
+        requested_size: u64,
+        /// Buffer type (which determines the [max size](BufferType::max_size))
+        buffer_type: BufferType,
+    },
+}
+
+/// Buffer type for error reporting
+#[derive(Copy, Clone, Debug)]
+pub enum BufferType {
+    /// Uniform buffer ([`wgpu::BufferUsages::UNIFORM`])
+    Uniform,
+    /// Storage buffer ([`wgpu::BufferUsages::STORAGE`])
+    Storage,
+    /// Other buffer type (e.g. [`wgpu::BufferUsages::MAP_READ`])
+    Generic,
+}
+
+impl std::fmt::Display for BufferType {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let s = match self {
+            BufferType::Uniform => "uniform",
+            BufferType::Storage => "storage",
+            BufferType::Generic => "generic",
+        };
+        s.fmt(f)
+    }
+}
+
+impl BufferType {
+    /// Maximum size of this buffer type, per the WebGPU spec
+    pub const fn max_size(&self) -> u64 {
+        // These are copied from the spec, since we don't ask for anything extra
+        match self {
+            // maxUniformBufferBindingSize
+            BufferType::Uniform => 64 * 1024,
+            // maxStorageBufferBindingSize
+            BufferType::Storage => 128 * 1024 * 1024,
+            // maxBufferSize
+            BufferType::Generic => 256 * 1024 * 1024,
+        }
+    }
+
+    fn check(&self, requested_size: u64) -> Result<(), BufferSizeError> {
+        if !requested_size.is_multiple_of(4) {
+            Err(BufferSizeError::NotAligned(requested_size))
+        } else if requested_size > self.max_size() {
+            Err(BufferSizeError::TooLarge {
+                requested_size,
+                buffer_type: *self,
+            })
+        } else {
+            Ok(())
+        }
+    }
 }
